% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{balanced_accuracy_score}
\alias{balanced_accuracy_score}
\title{Computes balanced accuracy.}
\usage{
balanced_accuracy_score(true_labels, predicted_labels)
}
\arguments{
\item{true_labels}{The true class labels.}

\item{predicted_labels}{The predicted class labels.}
}
\value{
The balanced accuracy score.
}
\description{
This function requires the 'caret' package.
}
