% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_functions.R
\name{curvature_estimation}
\alias{curvature_estimation}
\title{Computes the curvatures of all samples in the training set.}
\usage{
curvature_estimation(data, k)
}
\arguments{
\item{data}{A numeric matrix or data frame of the training data.}

\item{k}{The number of neighbors for the initial k-NN graph.}
}
\value{
A numeric vector of curvatures for each sample.
}
\description{
Computes the curvatures of all samples in the training set.
}
