% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f1_score}
\alias{f1_score}
\title{Computes the F1-score.}
\usage{
f1_score(true_labels, predicted_labels, average = "weighted")
}
\arguments{
\item{true_labels}{The true class labels.}

\item{predicted_labels}{The predicted class labels.}

\item{average}{The type of averaging ('weighted').}
}
\value{
The F1-score.
}
\description{
Computes the F1-score.
}
