% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_functions.R
\name{point_curvature_estimation}
\alias{point_curvature_estimation}
\title{Computes the curvature of a single test sample's neighborhood.}
\usage{
point_curvature_estimation(data)
}
\arguments{
\item{data}{A numeric matrix or data frame representing the neighborhood (test point + its neighbors).}
}
\value{
A single numeric value for the curvature.
}
\description{
Computes the curvature of a single test sample's neighborhood.
}
