% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature_functions.R
\name{quantize}
\alias{quantize}
\title{Quantizes real values to integer levels.}
\usage{
quantize(arr, a, b, k = 10)
}
\arguments{
\item{arr}{A numeric vector in the interval \code{[a, b]}.}

\item{a}{The lower bound of the interval.}

\item{b}{The upper bound of the interval.}

\item{k}{The number of quantization levels (default is 10).}
}
\value{
A vector of quantized integers in \code{0, ..., k - 1}.
}
\description{
This function quantizes real values in the interval \code{[a, b]} to integer levels from 0 to k-1.
}
