% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testa_KNN}
\alias{testa_KNN}
\title{Standard k-NN classifier.}
\usage{
testa_KNN(train, test, train_target, nn)
}
\arguments{
\item{train}{A numeric matrix or data frame of the training data.}

\item{test}{A numeric matrix or data frame of the test data.}

\item{train_target}{A numeric or factor vector of class labels for the training data.}

\item{nn}{The number of neighbors.}
}
\value{
A numeric or factor vector of predicted class labels.
}
\description{
Standard k-NN classifier.
}
